CHORE: Improve test efficiency (#6196)
