+++ doc/changelog.d/6196.maintenance.md
@@ -1,0 +1,1 @@
+Improve test efficiency
