+++ tests/system/general/test_98_Icepak.py
@@ -31,7 +31,6 @@ from ansys.aedt.core import Icepak
 from ansys.aedt.core.generic.settings import settings
 from ansys.aedt.core.internal.errors import AEDTRuntimeError
 from ansys.aedt.core.modules.boundary.icepak_boundary import NetworkObject
-from ansys.aedt.core.modules.boundary.layout_boundary import NativeComponentObject
 from ansys.aedt.core.modules.boundary.layout_boundary import PCBSettingsDeviceParts
 from ansys.aedt.core.modules.boundary.layout_boundary import PCBSettingsPackageParts
 from ansys.aedt.core.modules.mesh_icepak import MeshRegion
@@ -65,13 +64,6 @@ en_PreserveResults = True
 link_data = [proj_name, design_name, solution_name, en_ForceSimulation, en_PreserveResults]
 solution_freq = "2.5GHz"
 resolution = 2
-
-
-@pytest.fixture(scope="class", autouse=True)
-def dummy_prj(add_app):
-    app = add_app("Dummy_license_checkout_prj")
-    yield app
-    app.close_project(app.project_name)
 
 
 @pytest.fixture(autouse=True)
@@ -696,17 +688,18 @@ class TestClass:
         assert ipk.native_components["Fan1"].props["NativeComponentDefinitionProvider"]["Swirl"] == "10"
 
     def test035__create_heat_sink(self, ipk):
-        assert ipk.create_parametric_fin_heat_sink(
-            draftangle=1.5,
-            patternangle=8,
-            numcolumn_perside=3,
-            vertical_separation=5.5,
-            material="Copper",
-            center=[10, 0, 0],
-            plane_enum=ipk.PLANE.XY,
-            rotation=45,
-            tolerance=0.005,
-        )
+        # Deprecated many versions ago
+        # assert ipk.create_parametric_fin_heat_sink(
+        #     draftangle=1.5,
+        #     patternangle=8,
+        #     numcolumn_perside=3,
+        #     vertical_separation=5.5,
+        #     material="Copper",
+        #     center=[10, 0, 0],
+        #     plane_enum=ipk.PLANE.XY,
+        #     rotation=45,
+        #     tolerance=0.005,
+        # )
         box = ipk.modeler.create_box([0, 0, 0], [20, 20, 3])
         top_face = box.top_face_z
         hs, _ = ipk.create_parametric_heatsink_on_face(top_face, material="Al-Extruded")
@@ -866,86 +859,6 @@ class TestClass:
 
     @pytest.mark.skipif(not config["use_grpc"], reason="Not running in COM mode")
     @pytest.mark.parametrize("ipk", [board_ipk], indirect=True)
-    def test047__advanced3dcomp_export(self, ipk, local_scratch):
-        surf1 = ipk.modeler.create_rectangle(ipk.PLANE.XY, [0, 0, 0], [10, 20], name="surf1")
-        box1 = ipk.modeler.create_box([20, 20, 2], [10, 10, 3], "box1", "copper")
-        fan = ipk.create_fan("Fan", cross_section="YZ", radius="1mm", hub_radius="0mm")
-        cs0 = ipk.modeler.create_coordinate_system(name="CS0")
-        cs0.props["OriginX"] = 10
-        cs0.props["OriginY"] = 10
-        cs0.props["OriginZ"] = 10
-        cs1 = ipk.modeler.create_coordinate_system(name="CS1", reference_cs="CS0")
-        cs1.props["OriginX"] = 10
-        cs1.props["OriginY"] = 10
-        cs1.props["OriginZ"] = 10
-        fan2_prop = dict(fan.props).copy()
-        fan2_prop["TargetCS"] = "CS1"
-        fan2 = NativeComponentObject(ipk, "Fan", "Fan2", fan2_prop)
-        fan2.create()
-        ipk.create_ipk_3dcomponent_pcb(
-            "Board", link_data, solution_freq, resolution, custom_x_resolution=400, custom_y_resolution=500
-        )
-        ipk.monitor.assign_face_monitor(box1.faces[0].id, "Temperature", "FaceMonitor")
-        ipk.monitor.assign_point_monitor_in_object(box1.name, "Temperature", "BoxMonitor")
-        ipk.monitor.assign_surface_monitor(surf1.name, "Temperature", "SurfaceMonitor")
-        ipk.create_dataset(
-            "test_dataset",
-            [1, 2, 3, 4],
-            [1, 2, 3, 4],
-            z=None,
-            v=None,
-            is_project_dataset=False,
-            x_unit="cel",
-            y_unit="W",
-            v_unit="",
-        )
-        file_path = local_scratch.path
-        file_name = "Advanced3DComp.a3dcomp"
-        fan_obj = ipk.create_fan(is_2d=True)
-        ipk.monitor.assign_surface_monitor(
-            list(ipk.modeler.user_defined_components[fan_obj.name].parts.values())[0].name
-        )
-        ipk.monitor.assign_face_monitor(
-            list(ipk.modeler.user_defined_components[fan_obj.name].parts.values())[0].faces[0].id
-        )
-        fan_obj_3d = ipk.create_fan(is_2d=False)
-        ipk.monitor.assign_point_monitor_in_object(
-            list(ipk.modeler.user_defined_components[fan_obj_3d.name].parts.values())[0].name
-        )
-        assert ipk.modeler.create_3dcomponent(
-            os.path.join(file_path, file_name),
-            name="board_assembly",
-            coordinate_systems=["Global"],
-            export_auxiliary=True,
-        )
-        ipk.create_dataset(
-            "test_ignore",
-            [1, 2, 3, 4],
-            [1, 2, 3, 4],
-            z=None,
-            v=None,
-            is_project_dataset=False,
-            x_unit="cel",
-            y_unit="W",
-            v_unit="",
-        )
-        file_name = "Advanced3DComp1.a3dcomp"
-        mon_list = list(ipk.monitor.all_monitors.keys())
-        ipk.monitor.assign_point_monitor([0, 0, 0])
-        cs_list = [cs.name for cs in ipk.modeler.coordinate_systems if cs.name != "CS0"]
-        ipk.modeler.create_coordinate_system()
-        assert ipk.modeler.create_3dcomponent(
-            os.path.join(file_path, file_name),
-            name="board_assembly",
-            coordinate_systems=cs_list,
-            reference_coordinate_system="CS1",
-            export_auxiliary=True,
-            monitor_objects=mon_list,
-            datasets=["test_dataset"],
-        )
-
-    @pytest.mark.skipif(not config["use_grpc"], reason="Not running in COM mode")
-    @pytest.mark.parametrize("ipk", [board_ipk], indirect=True)
     def test048__advanced3dcomp_import(self, ipk, local_scratch):
         cs2 = ipk.modeler.create_coordinate_system(name="CS2")
         cs2.props["OriginX"] = 20
@@ -976,73 +889,6 @@ class TestClass:
             coordinate_system="Global",
             name="test",
             auxiliary_parameters=False,
-        )
-
-    @pytest.mark.skipif(not config["use_grpc"], reason="Not running in COM mode")
-    def test049__flatten_3d_components(self, ipk, local_scratch):
-        file_path = local_scratch.path
-        file_name = "Advanced3DComp_T52.a3dcomp"
-        ipk.create_fan(is_2d=True)
-        box = ipk.modeler.create_box([5, 6, 7], [9, 7, 6])
-        rectangle = ipk.modeler.create_rectangle(0, [5, 6, 7], [7, 6])
-        ipk.monitor.assign_surface_monitor(rectangle.name)
-        ipk.monitor.assign_face_monitor(box.faces[0].id)
-        ipk.create_fan(is_2d=False)
-        ipk.monitor.assign_point_monitor_in_object(box.name)
-        ipk.create_dataset(
-            "test_ignore",
-            [1, 2, 3, 4],
-            [1, 2, 3, 4],
-            z=None,
-            v=None,
-            is_project_dataset=False,
-            x_unit="cel",
-            y_unit="W",
-            v_unit="",
-        )
-        mon_list = list(ipk.monitor.all_monitors.keys())
-        ipk.monitor.assign_point_monitor([0, 0, 0])
-        ipk.modeler.create_coordinate_system()
-        assert ipk.modeler.create_3dcomponent(
-            os.path.join(file_path, file_name),
-            name="board_assembly",
-            coordinate_systems=["Global"],
-            reference_coordinate_system="Global",
-            export_auxiliary=True,
-            monitor_objects=mon_list,
-            datasets=["test_dataset"],
-        )
-        ipk.insert_design("test_52_1")
-        cs2 = ipk.modeler.create_coordinate_system(name="CS2")
-        cs2.props["OriginX"] = 20
-        cs2.props["OriginY"] = 20
-        cs2.props["OriginZ"] = 20
-        file_path = local_scratch.path
-        ipk.modeler.insert_3d_component(
-            input_file=os.path.join(file_path, file_name), coordinate_system="CS2", auxiliary_parameters=True
-        )
-
-        ipk.logger.clear_messages("", "")
-        ipk.insert_design("test_52")
-        cs2 = ipk.modeler.create_coordinate_system(name="CS2")
-        cs2.props["OriginX"] = 20
-        cs2.props["OriginY"] = 20
-        cs2.props["OriginZ"] = 20
-        file_path = local_scratch.path
-        ipk.modeler.insert_3d_component(
-            input_file=os.path.join(file_path, file_name), coordinate_system="CS2", auxiliary_parameters=True
-        )
-        mon_name = ipk.monitor.assign_face_monitor(
-            list(ipk.modeler.user_defined_components["board_assembly1"].parts.values())[0].faces[0].id
-        )
-        mon_point_name = ipk.monitor.assign_point_monitor([20, 20, 20])
-        assert ipk.flatten_3d_components()
-        assert all(
-            i in ipk.monitor.all_monitors
-            for i in [
-                mon_name,
-                mon_point_name,
-            ]
         )
 
     def test050__create_conduting_plate(self, ipk):
